+++ lib/mplcairo/gtk.py
@@ -1,7 +1,5 @@
-import cairo
 from matplotlib.backends.backend_gtk3 import _BackendGTK3, FigureCanvasGTK3
 
-from . import _util
 from .base import FigureCanvasCairo
 
 
@@ -13,12 +11,10 @@ class FigureCanvasGTKCairo(FigureCanvasCairo, FigureCanvasGTK3):
         # We always repaint the full canvas (doing otherwise would require an
         # additional copy of the buffer into a contiguous block, so it's not
         # clear it would be faster).
-        buf = _util.cairo_to_premultiplied_argb32(
-            self.get_renderer(_ensure_drawn=True)._get_buffer())
-        height, width, _ = buf.shape
-        image = cairo.ImageSurface.create_for_data(
-            buf, cairo.FORMAT_ARGB32, width, height)
-        ctx.set_source_surface(image, 0, 0)
+        surface = \
+            self.get_renderer(_ensure_drawn=True)._get_context().get_target()
+        surface.flush()
+        ctx.set_source_surface(surface, 0, 0)
         ctx.paint()
 
     def blit(self, bbox=None):  # FIXME: flickering.